requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
Deprecated==1.2.10
discord.py==1.6.0a2699+gede7b47
idna==2.10
jsons==1.3.0
multidict==5.1.0
PyGithub==1.54
PyJWT==1.7.1
requests==2.24.0
typing-extensions==3.7.4.3
typish==1.9.1
urllib3==1.25.11
wrapt==1.12.1
yarl==1.6.3
